Tier 1 refactoring: extract shared utilities to eliminate duplication
- lib/format.tsx: single source for fmtDuration, fmt, stripHtml, timeAgo,
  linkify (with configurable link class), extractImages, stripNostrUris,
  fireConfetti — removes ~180 lines of copy-paste across 8 components
- lib/pubsub.ts: createObservable() replaces identical listeners/notify/
  subscribe boilerplate in nwc.ts, webln.ts, and spark.ts
- lib/nostr/event-queries.ts: fetchLatestEvent() replaces pool.querySync +
  sort + take-first boilerplate in favorites.ts, mutes.ts, profile.ts,
  and wallet-backup.ts (~40 lines of boilerplate eliminated)
