+++ components/player.tsx
@@ -1,18 +1,10 @@
 'use client';
 import { useEffect, useRef, useState } from 'react';
 import { useApp } from '@/lib/store';
+import { fmt } from '@/lib/format';
 import { BoostModal } from './boost-modal';
 import { BoltIcon } from './icons';
 import { FullscreenPlayer } from './fullscreen-player';
-
-function fmt(t: number) {
-  if (!isFinite(t)) return '0:00';
-  const h = Math.floor(t / 3600);
-  const m = Math.floor((t % 3600) / 60);
-  const s = Math.floor(t % 60).toString().padStart(2, '0');
-  if (h > 0) return `${h}:${m.toString().padStart(2, '0')}:${s}`;
-  return `${m}:${s}`;
-}
 
 export function Player() {
   const { current, isPlaying, setPlaying, setPosition, positionSec } = useApp();
